+++ client/app/src/main/java/com/microsoft/research/karya/ui/payment/dashboard/PaymentDashboardViewModel.kt
@@ -73,7 +73,7 @@ constructor(
 
           val userTransactionDetail =
             UserTransactionDetail(
-              amount = transaction.amount.toFloat(),
+              amount = transaction.amount,
               utr = transaction.utr,
               date = userDate ?: "",
               status = transaction.status
